+++ test/server/main.cpp
@@ -5,12 +5,6 @@
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 //
 
-#ifndef BEAST_EXAMPLE_HTTP_CRAWL_URLS_LARGE_DATA_HPP
-#define BEAST_EXAMPLE_HTTP_CRAWL_URLS_LARGE_DATA_HPP
-
-#include <vector>
-
-std::vector<char const*> const&
-urls_large_data();
-
-#endif
+int main(int argc, char** argv)
+{
+}