+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 60:
 
 * String comparisons are public interfaces
 * Fix response message type in async websocket accept
+* New server-framework, full featured server example
 
 --------------------------------------------------------------------------------